Seperating webhook connection from http client connection; adding test for webhook connection
