+++ tests/test_examples/test_http_client_connection_to_aries_cloud_agent.py
@@ -81,6 +81,7 @@ class TestAEAToACA:
             )
 
         # run an ACA
+        # command: aca-py start --admin 127.0.0.1 8020 --admin-insecure-mode --inbound-transport http 0.0.0.0 8000 --outbound-transport http
         cls.process = subprocess.Popen(  # nosec
             [
                 "aca-py",
